# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Kirana Store Manager API — Multi-tenant SaaS
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /items:
    get:
      operationId: listItems
      tags: [inventory]
      summary: List all inventory items
      parameters:
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: lowStock
          schema:
            type: string
        - in: query
          name: q
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Item"
    post:
      operationId: createItem
      tags: [inventory]
      summary: Create a new inventory item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateItemBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
  /items/{id}:
    get:
      operationId: getItem
      tags: [inventory]
      summary: Get item by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
        "404":
          description: Not found
    patch:
      operationId: updateItem
      tags: [inventory]
      summary: Update an inventory item
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateItemBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
    delete:
      operationId: deleteItem
      tags: [inventory]
      summary: Delete an inventory item
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
  /inventory/default-items:
    get:
      operationId: getDefaultItems
      tags: [inventory]
      summary: Get default items catalog
      parameters:
        - in: query
          name: q
          schema:
            type: string
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of default items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Item"
  /invoices:
    get:
      operationId: listInvoices
      tags: [invoices]
      summary: List all invoices
      parameters:
        - in: query
          name: startDate
          schema:
            type: string
        - in: query
          name: endDate
          schema:
            type: string
        - in: query
          name: paymentMethod
          schema:
            type: string
            enum: [cash, upi]
        - in: query
          name: minAmount
          schema:
            type: number
        - in: query
          name: maxAmount
          schema:
            type: number
        - in: query
          name: q
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of invoices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invoice"
    post:
      operationId: createInvoice
      tags: [invoices]
      summary: Create a new invoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvoiceBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /invoices/{id}:
    get:
      operationId: getInvoice
      tags: [invoices]
      summary: Get invoice by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
        "404":
          description: Not found
  /invoices/{id}/pdf:
    get:
      operationId: getInvoicePdf
      tags: [invoices]
      summary: Download invoice PDF
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: PDF file
          content:
            application/pdf:
              schema:
                type: string
                format: binary
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List all customers
      parameters:
        - in: query
          name: q
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create a new customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomerBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get customer by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "404":
          description: Not found
    patch:
      operationId: updateCustomer
      tags: [customers]
      summary: Update a customer
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCustomerBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    delete:
      operationId: deleteCustomer
      tags: [customers]
      summary: Delete a customer
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
  /ai/ocr:
    post:
      operationId: processOcr
      tags: [ai]
      summary: Process image with OCR to extract invoice items
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OcrRequest"
      responses:
        "200":
          description: Extracted items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiExtractResponse"
  /ai/voice:
    post:
      operationId: processVoice
      tags: [ai]
      summary: Process voice/text input to extract invoice items
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VoiceRequest"
      responses:
        "200":
          description: Extracted items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiExtractResponse"
components:
  schemas:
...
    Item:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        category:
          type: string
        price:
          type: number
        stock:
          type: number
        unit:
          type: string
          enum: [kg, pcs, litre, gm, ml, dozen, pack]
        lowStockThreshold:
          type: number
        isLowStock:
          type: boolean
        createdAt:
          type: string
      required: [id, name, category, price, stock, unit, lowStockThreshold, isLowStock, createdAt]
    CreateItemBody:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        price:
          type: number
        stock:
          type: number
        unit:
          type: string
          enum: [kg, pcs, litre, gm, ml, dozen, pack]
        lowStockThreshold:
          type: number
      required: [name, category, price, stock, unit]
    UpdateItemBody:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        price:
          type: number
        stock:
          type: number
        unit:
          type: string
          enum: [kg, pcs, litre, gm, ml, dozen, pack]
        lowStockThreshold:
          type: number
...
    InvoiceItem:
      type: object
      properties:
        itemId:
          type: string
        itemName:
          type: string
        quantity:
          type: number
        unit:
          type: string
        price:
          type: number
        total:
          type: number
      required: [itemId, itemName, quantity, unit, price, total]
    Invoice:
      type: object
      properties:
        id:
          type: string
        invoiceNumber:
          type: string
        customerName:
          type: ["string", "null"]
        customerPhone:
          type: ["string", "null"]
        customerId:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceItem"
        subtotal:
          type: number
        total:
          type: number
        paymentMethod:
          type: string
          enum: [cash, upi]
        status:
          type: string
          enum: [paid, unpaid]
        pdfUrl:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, invoiceNumber, items, subtotal, total, paymentMethod, status, createdAt]
    CreateInvoiceBody:
      type: object
      properties:
        customerName:
          type: ["string", "null"]
        customerPhone:
          type: ["string", "null"]
        customerId:
          type: ["string", "null"]
        items:
          type: array
          items:
            type: object
            properties:
              itemId:
                type: string
              quantity:
                type: number
            required: [itemId, quantity]
        paymentMethod:
          type: string
          enum: [cash, upi]
        status:
          type: string
          enum: [paid, unpaid]
      required: [items, paymentMethod]
    Customer:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, name, createdAt]
    CreateCustomerBody:
      type: object
      properties:
        name:
          type: string
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
      required: [name]
    UpdateCustomerBody:
      type: object
      properties:
        name:
          type: string
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
...
    AiExtractedItem:
      type: object
      properties:
        name:
          type: string
        quantity:
          type: number
        unit:
          type: string
        matchedItemId:
          type: ["string", "null"]
        matchedItemName:
          type: ["string", "null"]
        confidence:
          type: number
      required: [name, quantity, unit, confidence]
    AiExtractResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AiExtractedItem"
        rawText:
          type: string
        success:
          type: boolean
      required: [items, rawText, success]
    OcrRequest:
      type: object
      properties:
        imageBase64:
          type: string
      required: [imageBase64]
    VoiceRequest:
      type: object
      properties:
        text:
          type: string
      required: [text]